Can now read and expose the contents of an ISO
This PR enables `plunder` to use the content of an ISO without mounting it, removing the need for any elevated permissions. Prior to this a user would need to loopback mount an ISO and use that path as a boot mirror for the installation, with this in place we use a `isoPrefix` identifier and a path to expose the ISO as a content store to Plunder.